// src/lib/api/schema.gen.ts
export interface components {
    schemas: {
// ...
        Snapshot: {
            /** @enum {number} */
            schemaVersion: 1;
            beans: components["schemas"]["Bean"][];
            machines: components["schemas"]["Machine"][];
            brews: components["schemas"]["Brew"][];
        };
        SnapshotResponse: {
            snapshot: components["schemas"]["Snapshot"];
            /**
             * @description Current snapshot version for concurrency control
             * @example 0
             */
            version: number;
        };
// ...
    };
// ...
}
// ...
type FlattenedDeepRequired<T> = {
    [K in keyof T]-?: FlattenedDeepRequired<T[K] extends unknown[] | undefined | null ? Extract<T[K], unknown[]>[number] : T[K]>;
};
type ReadonlyArray<T> = [
    Exclude<T, undefined>
] extends [
    unknown[]
] ? Readonly<Exclude<T, undefined>> : Readonly<Exclude<T, undefined>[]>;
// ...
export const beanStatusValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["Bean"]["status"]> = ["Excellent", "Good", "Mid", "Horrible", "New", ""];
export const beanDominantNoteValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["Bean"]["dominantNote"]> = ["Fruity", "Nutty", "Floral", "Sweet", "Sour", "Spices", "Roasted", "Green", ""];
export const beanBotanicValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["Bean"]["botanic"]> = ["Arabica", "Robusta", ""];
export const beanDesignationValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["Bean"]["designation"]> = ["Pure Origin", "Blend", ""];